backend/requirements.txt
# FastAPI и основные зависимости
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# База данных
sqlalchemy>=2.0.23
alembic>=1.13.0
aiosqlite>=0.19.0

# Аутентификация и безопасность
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP клиент
httpx>=0.25.2
aiofiles>=23.2.1

# Redis
redis>=5.0.1
aioredis>=2.0.1

# Мониторинг
prometheus-client>=0.19.0
psutil>=5.9.6

# QR коды
qrcode[pil]>=7.4.2

# Логирование
structlog>=23.2.0
colorama>=0.4.6

# Утилиты
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0

# Криптография для VPN ключей
cryptography>=41.0.7

# Валидация
email-validator>=2.1.0

# Работа с JSON и датами
orjson>=3.9.10
python-dateutil>=2.8.2

# WebSocket
websockets>=12.0

# Тестирование
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
